# mail-vue/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cloud Mail API",
    "version": "1.0.0",
    "description": "REST API for a user to manage only their own mail accounts, messages, and attachments. API access must be enabled by the administrator. Create and revoke named API keys from the authenticated web settings page."
  },
  "servers": [
    {
      "url": "https://mail.airoute.kdns.fr/api/v1",
      "description": "Production"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "tags": [
    {
      "name": "Current user"
    },
    {
      "name": "Accounts"
    },
    {
      "name": "Emails"
    },
    {
      "name": "Attachments"
    }
  ],
  "paths": {
    "/me": {
      "get": {
        "tags": [
          "Current user"
        ],
        "summary": "Get the API key owner",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/accounts": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "List the current user's mail accounts",
        "operationId": "listAccounts",
        "parameters": [
          {
            "$ref": "#/components/parameters/ForwardCursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Account page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Create a child mail account",
        "description": "The account is always owned by the API key owner. Domain, prefix, uniqueness, role domain, and account-count restrictions still apply. Interactive Turnstile verification is not required after API key authentication.",
        "operationId": "createAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/accounts/{accountId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AccountId"
        }
      ],
      "patch": {
        "tags": [
          "Accounts"
        ],
        "summary": "Update an owned mail account",
        "operationId": "updateAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Delete an owned child mail account",
        "description": "The user's primary account cannot be deleted.",
        "operationId": "deleteAccount",
        "responses": {
          "200": {
            "$ref": "#/components/responses/EmptySuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/emails": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "List the current user's email summaries",
        "description": "Uses a descending email ID cursor. The response deliberately excludes full HTML and text bodies.",
        "operationId": "listEmails",
        "parameters": [
          {
            "$ref": "#/components/parameters/BackwardCursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "description": "Filter by an account owned by the current user.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "description": "0 = received, 1 = sent.",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            }
          },
          {
            "name": "unread",
            "in": "query",
            "required": false,
            "description": "0 = unread, 1 = read.",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailPageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/emails/{emailId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/EmailId"
        }
      ],
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get one owned email including its body",
        "operationId": "getEmail",
        "responses": {
          "200": {
            "description": "Email details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDetailResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Emails"
        ],
        "summary": "Soft-delete one owned email",
        "operationId": "deleteEmail",
        "responses": {
          "200": {
            "$ref": "#/components/responses/EmptySuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/emails/{emailId}/read": {
      "parameters": [
        {
          "$ref": "#/components/parameters/EmailId"
        }
      ],
      "patch": {
        "tags": [
          "Emails"
        ],
        "summary": "Mark one owned email as read",
        "operationId": "markEmailRead",
        "responses": {
          "200": {
            "$ref": "#/components/responses/EmptySuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/emails/{emailId}/attachments": {
      "parameters": [
        {
          "$ref": "#/components/parameters/EmailId"
        }
      ],
      "get": {
        "tags": [
          "Attachments"
        ],
        "summary": "List metadata for attachments on an owned email",
        "operationId": "listAttachments",
        "responses": {
          "200": {
            "description": "Attachment metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/emails/{emailId}/attachments/{attachmentId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/EmailId"
        },
        {
          "$ref": "#/components/parameters/AttachmentId"
        }
      ],
      "get": {
        "tags": [
          "Attachments"
        ],
        "summary": "Download one attachment after ownership verification",
        "operationId": "downloadAttachment",
        "responses": {
          "200": {
            "description": "Attachment bytes",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "contentEncoding": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "cma_publicId_secret",
        "description": "Create a named key in the authenticated user settings page. Send it as `Authorization: Bearer cma_...`."
      }
    },
    "parameters": {
      "AccountId": {
        "name": "accountId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "EmailId": {
        "name": "emailId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "AttachmentId": {
        "name": "attachmentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "ForwardCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Return resources with an ID greater than this cursor.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "BackwardCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Return email IDs lower than this cursor.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 50,
          "default": 20
        }
      }
    },
    "responses": {
      "EmptySuccess": {
        "description": "Operation completed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptySuccessResponse"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, malformed, unknown, or revoked API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "REST API disabled, API key owner disabled, or operation forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Owned resource not found. Cross-user IDs are also reported as not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource conflict or protected primary account",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "CurrentUser": {
        "type": "object",
        "required": [
          "userId",
          "email",
          "name",
          "type"
        ],
        "properties": {
          "userId": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "integer"
          },
          "role": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "Account": {
        "type": "object",
        "required": [
          "accountId",
          "email",
          "name",
          "status",
          "allReceive",
          "sort"
        ],
        "properties": {
          "accountId": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "latestEmailTime": {
            "type": [
              "string",
              "null"
            ]
          },
          "createTime": {
            "type": [
              "string",
              "null"
            ]
          },
          "allReceive": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "sort": {
            "type": "integer"
          }
        }
      },
      "CreateAccountRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "UpdateAccountRequest": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 30
          },
          "allReceive": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          }
        }
      },
      "EmailSummary": {
        "type": "object",
        "properties": {
          "emailId": {
            "type": "integer"
          },
          "accountId": {
            "type": "integer"
          },
          "sendEmail": {
            "type": [
              "string",
              "null"
            ]
          },
          "sendName": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject": {
            "type": [
              "string",
              "null"
            ]
          },
          "toEmail": {
            "type": "string"
          },
          "toName": {
            "type": "string"
          },
          "type": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "status": {
            "type": "integer"
          },
          "unread": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "code": {
            "type": "string"
          },
          "textPreview": {
            "type": "string"
          },
          "createTime": {
            "type": "string"
          }
        }
      },
      "EmailDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EmailSummary"
          },
          {
            "type": "object",
            "properties": {
              "text": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "content": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "recipient": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "cc": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "bcc": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "messageId": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "inReplyTo": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "attachmentCount": {
                "type": "integer",
                "minimum": 0
              }
            }
          }
        ]
      },
      "Attachment": {
        "type": "object",
        "properties": {
          "attachmentId": {
            "type": "integer"
          },
          "emailId": {
            "type": "integer"
          },
          "accountId": {
            "type": "integer"
          },
          "filename": {
            "type": [
              "string",
              "null"
            ]
          },
          "mimeType": {
            "type": [
              "string",
              "null"
            ]
          },
          "size": {
            "type": [
              "integer",
              "null"
            ]
          },
          "disposition": {
            "type": [
              "string",
              "null"
            ]
          },
          "contentId": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": "integer"
          },
          "createTime": {
            "type": "string"
          }
        }
      },
      "CurrentUserResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CurrentUser"
          }
        }
      },
      "AccountResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Account"
          }
        }
      },
      "AccountPageResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "items",
              "nextCursor"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Account"
                }
              },
              "nextCursor": {
                "type": [
                  "integer",
                  "null"
                ]
              }
            }
          }
        }
      },
      "EmailPageResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "items",
              "nextCursor"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EmailSummary"
                }
              },
              "nextCursor": {
                "type": [
                  "integer",
                  "null"
                ]
              }
            }
          }
        }
      },
      "EmailDetailResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EmailDetail"
          }
        }
      },
      "AttachmentListResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          }
        }
      },
      "EmptySuccessResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "null"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "status",
              "code",
              "message"
            ],
            "properties": {
              "status": {
                "type": "integer"
              },
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}